requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
qrcode[pil]==7.4.2
python-barcode[images]==0.15.1
pillow==10.1.0
python-multipart==0.0.6
aiofiles==23.2.1
apscheduler==3.10.4
requests==2.31.0
python-dotenv==1.0.0